+++ backend/src/app/core/db/elasticsearch_service.py
