+++ backend/src/app/core/search/sdoc_search/sdoc_search.py
@@ -12,6 +12,8 @@ from app.core.data.dto.search import (
     SimSearchSentenceHit,
 )
 from app.core.data.orm.source_document import SourceDocumentORM
+from app.core.db.elasticsearch_service import ElasticSearchService
+from app.core.db.simsearch_service import SimSearchService
 from app.core.db.sql_service import SQLService
 from app.core.filters.column_info import ColumnInfo
 from app.core.filters.filtering import (
@@ -19,9 +21,7 @@ from app.core.filters.filtering import (
 )
 from app.core.filters.search_builder import SearchBuilder
 from app.core.filters.sorting import Sort
-from app.core.search.elasticsearch_service import ElasticSearchService
 from app.core.search.sdoc_search.sdoc_search_columns import SearchColumns
-from app.core.search.simsearch_service import SimSearchService
 
 
 def search_info(project_id) -> List[ColumnInfo[SearchColumns]]:
