+++ backend/src/app/preprocessing/pipeline/steps/text/store_document_in_elasticsearch.py
@@ -1,7 +1,7 @@
 from app.core.data.dto.search import (
     ElasticSearchDocumentCreate,
 )
-from app.core.search.elasticsearch_service import ElasticSearchService
+from app.core.db.elasticsearch_service import ElasticSearchService
 from app.preprocessing.pipeline.model.pipeline_cargo import PipelineCargo
 from app.preprocessing.pipeline.model.text.preprotextdoc import PreProTextDoc
